+++ src/aks-preview/azext_aks_preview/_help.py
@@ -3185,7 +3185,7 @@ helps['aks machine add'] = """
        - name: --eviction-policy
          type: string
          short-summary: The eviction policy for machine. This cannot be specified unless the priority is 'Spot'. If not specified, the default is 'Delete'.
-       - name: --crg-id
+       - name: --capacity-reservation-group
          type: string
          short-summary: The fully qualified resource ID of the Capacity Reservation Group to associate with the machine.
 """